+++ splitcal/splitcal.cxx
@@ -87,7 +87,7 @@ Int_t splitcal::InitMedium(const char* name) {
     return -1111;
   }
   TGeoMedium* medium = gGeoManager->GetMedium(name);
-  if (medium != NULL) return ShipMedium->getMediumIndex();
+  if (medium != nullptr) return ShipMedium->getMediumIndex();
 
   return geoBuild->createMedium(ShipMedium);
 }